+++ packages/contracts/src/index.ts
@@ -4,6 +4,7 @@ export * from "./woocommerce";
 export * from "./ledger";
 export * from "./programme";
 export * from "./reporting";
+export * from "./experience";
 
 export const commerceEnvelopeV1 = z.object({
   version: z.literal("1"),